refactor(rig): split lease index lock
